refactor: move some files to maskbook package; make dependencies stricter
