+++ packages/dashboard/package.json
@@ -9,8 +9,11 @@
   },
   "dependencies": {
     "@dimensiondev/maskbook-shared": "*",
+    "@dimensiondev/maskbook-theme": "*",
+    "@dimensiondev/maskbook": "*",
     "color": "^3.1.3",
-    "@types/color": "^3.0.1"
+    "@types/color": "^3.0.1",
+    "async-call-rpc": "^5.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
