+++ packages/maskbook/package.json
@@ -2,11 +2,178 @@
   "name": "@dimensiondev/maskbook",
   "private": true,
   "license": "AGPL-3.0-or-later",
-  "scripts": {
-  },
+  "scripts": {},
   "dependencies": {
-    "@dimensiondev/maskbook-shared": "*"
+    "@dimensiondev/common-protocols": "1.6.0-20201027083702-d0ae6e2",
+    "@dimensiondev/holoflows-kit": "0.8.0-20201215084554-2f9e757",
+    "@dimensiondev/kit": "0.0.0-20210115033306-5d4304b",
+    "@dimensiondev/maskbook-shared": "*",
+    "@dimensiondev/maskbook-theme": "*",
+    "@dimensiondev/matrix-js-sdk-type": "=1.0.0-20200731085604-dc2de92",
+    "@dimensiondev/metamask-extension-provider": "^3.0.0-20210125122006-86e9bae",
+    "@dimensiondev/stego-js": "=0.11.1-20201027083223-8ab41be",
+    "@ethersproject/address": "^5.0.4",
+    "@ethersproject/contracts": "^5.0.4",
+    "@ethersproject/networks": "^5.0.4",
+    "@ethersproject/providers": "^5.0.9",
+    "@ethersproject/solidity": "^5.0.4",
+    "@msgpack/msgpack": "^2.3.0",
+    "@openzeppelin/contracts": "^3.2.0",
+    "@popperjs/core": "*",
+    "@servie/events": "^1.0.0",
+    "@types/bn.js": "^4.11.6",
+    "@types/classnames": "^2.2.11",
+    "@types/d3": "^6.2.0",
+    "@types/elliptic": "^6.4.12",
+    "@types/gun": "^0.9.2",
+    "@types/json-stable-stringify": "^1.0.32",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/lodash-es": "^4.17.4",
+    "@types/node": "^14.14.19",
+    "@types/qrcode": "^1.3.5",
+    "@types/react-highlight-words": "^0.16.1",
+    "@types/react-router-dom": "^5.1.6",
+    "@types/react-virtualized-auto-sizer": "^1.0.0",
+    "@types/react-window": "^1.8.2",
+    "@types/tiny-secp256k1": "^1.0.0",
+    "@types/use-subscription": "^1.0.0",
+    "@types/uuid": "^8.3.0",
+    "@ungap/promise-all-settled": "^1.1.2",
+    "@uniswap/sdk": "^3.0.3-beta.1",
+    "@walletconnect/client": "^1.2.2",
+    "anchorme": "^2.1.2",
+    "arweave": "^1.9.1",
+    "assert": "^2.0.0",
+    "async-call-rpc": "^5.0.0",
+    "bignumber.js": "^9.0.1",
+    "bip39": "^3.0.2",
+    "buffer": "^6.0.3",
+    "classnames": "^2.2.6",
+    "clipboard-polyfill": "^3.0.1",
+    "crypto-browserify": "^3.12.0",
+    "d3": "^5.16.0",
+    "date-fns": "^2.16.1",
+    "elliptic": "^6.5.3",
+    "eth-contract-metadata": "^1.15.0",
+    "ethereum-blockies": "^0.1.1",
+    "event-iterator": "^2.0.0",
+    "fuse.js": "^6.4.3",
+    "gun": "0.2020.520",
+    "https-browserify": "^1.0.0",
+    "i18next": "^19.8.4",
+    "idb": "^5.0.7",
+    "immer": "^8.0.1",
+    "json-stable-stringify": "^1.0.1",
+    "jsonwebtoken": "^8.5.1",
+    "jsqr": "^1.3.1",
+    "jss": "*",
+    "jsx-jsonml-devtools-renderer": "^1.4.3",
+    "lodash-es": "npm:lodash@^4.17.11",
+    "loglevel": "^1.7.0",
+    "matrix-js-sdk": "^9.3.0",
+    "next-tick": "^1.0.0",
+    "notistack": "^1.0.3",
+    "promievent": "^0.1.4",
+    "pvtsutils": "^1.0.14",
+    "qrcode": "^1.4.4",
+    "react-draggable": "^4.4.3",
+    "react-feather": "^2.0.9",
+    "react-highlight-words": "^0.16.0",
+    "react-i18next": "^11.8.4",
+    "react-middle-ellipsis": "^1.1.0",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-tilt": "^0.1.4",
+    "react-use": "^15.3.4",
+    "react-virtualized-auto-sizer": "^1.0.2",
+    "react-window": "^1.8.6",
+    "stream-browserify": "^3.0.0",
+    "stream-http": "^2.8.3",
+    "string.prototype.matchall": "^4.0.2",
+    "tiny-secp256k1": "^1.1.5",
+    "ts-results": "^3.0.0",
+    "use-subscription": "^1.5.1",
+    "uuid": "^8.3.2",
+    "wallet.ts": "^1.0.1",
+    "web-ext-types": "^3.1.0",
+    "web3": "^1.2.11",
+    "web3-core": "^1.2.11",
+    "web3-eth": "^1.3.0",
+    "web3-eth-contract": "^1.3.0",
+    "web3-eth-personal": "^1.3.0",
+    "web3-utils": "^1.3.0",
+    "webcrypto-liner": "1.2.3",
+    "webextension-polyfill": "^0.7.0",
+    "z-schema": "^5.0.0"
   },
   "devDependencies": {
+    "@dimensiondev/webextension-shim": "0.0.3-20201103060201-1bd22b9",
+    "@magic-works/commonjs-import.meta": "^1.0.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
+    "@sinonjs/text-encoding": "^0.7.1",
+    "@storybook/addon-actions": "^5.3.19",
+    "@storybook/addon-docs": "^5.3.19",
+    "@storybook/addon-info": "^5.3.20",
+    "@storybook/addon-knobs": "^5.3.19",
+    "@storybook/addon-links": "^5.3.19",
+    "@storybook/addon-viewport": "^5.3.19",
+    "@storybook/addons": "^5.3.17",
+    "@storybook/react": "^5.3.19",
+    "@testing-library/react-hooks": "^3.7.0",
+    "@types/elliptic": "^6.4.12",
+    "@types/enzyme": "^3.10.8",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@types/expect-puppeteer": "^4.4.5",
+    "@types/jest": "^26.0.19",
+    "@types/jest-environment-puppeteer": "^4.4.1",
+    "@types/puppeteer": "^5.4.2",
+    "@types/rimraf": "^3.0.0",
+    "@types/webpack-dev-server": "^3.11.1",
+    "@typescript-eslint/eslint-plugin": "^4.11.1-alpha.0",
+    "@typescript-eslint/parser": "^4.11.1",
+    "babel-loader": "^8.1.0",
+    "chrome-launcher": "^0.13.4",
+    "copy-webpack-plugin": "^7.0.0",
+    "dotenv": "^8.2.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
+    "expect-puppeteer": "^4.4.0",
+    "fake-indexeddb": "^3.1.2",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.0.0-beta.4",
+    "jest": "26.6.3",
+    "jest-environment-jsdom-fourteen": "^1.0.1",
+    "jest-esm-transformer": "^1.0.0",
+    "jest-junit": "^12.0.0",
+    "jest-puppeteer": "^4.4.0",
+    "jest-webextension-mock": "^3.7.6",
+    "openzeppelin-solidity": "^3.2.0",
+    "puppeteer": "^5.5.0",
+    "react-dev-utils": "^11.0.1",
+    "react-devtools-core": "^4.10.1",
+    "react-docgen-typescript-loader": "^3.7.2",
+    "react-refresh": "^0.9.0",
+    "react-refresh-typescript": "^1.1.0",
+    "react-test-renderer": "^17.0.1",
+    "rimraf": "^3.0.2",
+    "source-map-loader": "^1.1.3",
+    "storybook-addon-designs": "^5.4.2",
+    "storybook-addon-i18n": "^5.1.13",
+    "ts-generator": "^0.0.8",
+    "ts-jest": "26.4.4",
+    "ts-loader": "^8.0.13",
+    "typechain": "^4.0.1",
+    "typechain-target-web3-v1": "^1.0.4",
+    "web-ext": "^5.4.1",
+    "webpack": "^5.11.1",
+    "webpack-bundle-analyzer": "^4.3.0",
+    "webpack-cli": "^4.3.0",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-extension-manifest-plugin": "^0.6.0",
+    "webpack-notifier": "^1.12.0",
+    "webpack-target-webextension": "^0.4.0"
+  },
+  "peerDependencies": {
+    "jss": "*"
   }
 }
