+++ packages/maskbook/public/128x128.png
