+++ packages/maskbook/public/about-dialog-background.png
