+++ packages/maskbook/public/dashboard-placeholder-dark.png
