+++ packages/maskbook/public/dashboard-placeholder.png
