+++ packages/maskbook/public/image-payload/mask-transparent.png
