+++ packages/maskbook/public/image-payload/mask-v1.png
