+++ packages/maskbook/public/image-payload/mask-v2.png
