+++ packages/maskbook/public/image-payload/mask-v3.png
