+++ packages/maskbook/public/image-payload/mask-v4.png
