+++ packages/maskbook/public/image-payload/payload-v2.png
