+++ packages/maskbook/public/manifest.webextension-shim.json
