+++ packages/maskbook/public/restore-file-placeholder-dark.png
