+++ packages/maskbook/public/restore-file-placeholder-light.png
