+++ packages/maskbook/public/restore-image-placeholder-dark.png
