+++ packages/maskbook/public/restore-image-placeholder-light.png
