+++ packages/maskbook/public/wallet/payload-dai.png
