+++ packages/maskbook/public/wallet/payload-okb.png
