+++ packages/shared/package.json
@@ -1,6 +1,10 @@
 {
   "name": "@dimensiondev/maskbook-shared",
   "dependencies": {
+    "async-call-rpc": "^5.0.0",
+    "bignumber.js": "^9.0.1",
+    "immer": "^8.0.1",
+    "ts-results": "^3.1.1",
     "typeson": "^5.18.2",
     "typeson-registry": "^1.0.0-alpha.38"
   },
